simplify rtruedivide

+++ ndcube/conftest.py
@@ -469,7 +469,7 @@ def ndcube_2d_ln_lt(wcs_2d_lt_ln):
 @pytest.fixture
 def ndcube_2d_ln_lt_units(wcs_2d_lt_ln):
     shape = (10, 12)
-    data_cube = data_nd(shape)
+    data_cube = data_nd(shape).astype(float)
     return NDCube(data_cube, wcs=wcs_2d_lt_ln, unit=u.ct)
